create a local service worker that handles push notifications
