+++ frontend/dev.sh
@@ -10,6 +10,8 @@ cd openchat-shared
 pm2 start --name "shared" "npm run dev" 
 cd ../openchat-agent
 pm2 start --name "agent" "npm run dev" 
+cd ../openchat-push
+pm2 start --name "push" "npm run dev" 
 cd ../openchat-worker
 pm2 start --name "worker" "npm run dev" 
 cd ../openchat-client
