+++ frontend/openchat-push/.prettierignore
@@ -1,0 +1,1 @@
+**/candid/types.d.ts
