+++ frontend/openchat-push/.prettierrc
@@ -1,0 +1,6 @@
+{
+    "svelteBracketNewLine": false,
+    "tabWidth": 4,
+    "useTabs": false,
+    "printWidth": 100
+}
