+++ frontend/openchat-push/readme.md
@@ -1,0 +1,5 @@
+# OpenChat push notifications service worker
+
+This is a custom service worker designed to provide push notification services to OpenChat.
+
+Ideally this would be integrated with the default IC service worker, but for the time being that is not possible.
