+++ frontend/openchat-push/tsconfig.json
@@ -1,0 +1,26 @@
+{
+    "compilerOptions": {
+        "allowJs": true,
+        "checkJs": false,
+        "baseUrl": "./",
+        "declaration": true,
+        "experimentalDecorators": true,
+        "esModuleInterop": true,
+        "importsNotUsedAsValues": "error",
+        "lib": ["es2020", "WebWorker"],
+        "module": "es2020",
+        "moduleResolution": "node",
+        "noUnusedLocals": true,
+        "noUnusedParameters": true,
+        "outDir": "lib",
+        "paths": {},
+        "resolveJsonModule": true,
+        "rootDir": "./src",
+        "sourceMap": true,
+        "strict": true,
+        "target": "es2020",
+        "types": ["node"]
+    },
+    "exclude": ["node_modules/*", "../node_modules/*", "lib/*"],
+    "include": ["src/**/*"]
+}
