+++ frontend/openchat-sw/package.json
@@ -43,8 +43,6 @@
         "@dfinity/agent": "^0.14.0",
         "@dfinity/candid": "^0.14.0",
         "@dfinity/principal": "^0.14.0",
-        "openchat-shared": "*",
-        "openchat-agent": "*",
         "base64-arraybuffer": "^1.0.2",
         "pako": "^2.0.3",
         "siema": "^1.5.1",
